+++ drawing/src/bin/main_sc2.rs
@@ -9,8 +9,8 @@ use algebraeon_drawing::colour::Colour;
 use algebraeon_geometry::ambient_space::AffineSpace;
 use algebraeon_geometry::convex_hull::ConvexHull;
 use algebraeon_geometry::coordinates::Vector;
-use algebraeon_geometry::simplex::OrientationSide;
-use algebraeon_geometry::simplex::OrientedSimplex;
+use algebraeon_geometry::oriented_simplex::OrientationSide;
+use algebraeon_geometry::oriented_simplex::OrientedSimplex;
 use algebraeon_geometry::simplex_collection::LabelledSimplexCollection;
 use algebraeon_geometry::simplicial_disjoint_union::SimplicialDisjointUnion;
 use algebraeon_geometry::*;
